Add Markdown
For text areas in the app, there is now an option to edit & render
markdown. All #show pages with this feature, now have their 'Edit' and
'Back' buttons at the top of the page, rather than below the textarea.
I didn't do this on pages w/o markdown rendering. But later on, I may
want to make it consistent for all #show pages.